Move OBBSample to samples
