+++ src/super_gradients/training/samples/__init__.py
@@ -2,5 +2,6 @@ from .depth_estimation_sample import DepthEstimationSample
 from .pose_estimation_sample import PoseEstimationSample
 from .detection_sample import DetectionSample
 from .segmentation_sample import SegmentationSample
+from .obb_sample import OBBSample
 
-__all__ = ["PoseEstimationSample", "DetectionSample", "SegmentationSample", "DepthEstimationSample"]
+__all__ = ["PoseEstimationSample", "DetectionSample", "SegmentationSample", "DepthEstimationSample", "OBBSample"]
