+++ src/super_gradients/training/transforms/obb/obb_pad_if_needed.py
@@ -2,7 +2,7 @@ from typing import List
 
 from super_gradients.common.object_names import Processings
 from super_gradients.common.registry.registry import register_transform
-from .obb_sample import OBBSample
+from super_gradients.training.samples.obb_sample import OBBSample
 from super_gradients.training.transforms.utils import _pad_image, PaddingCoordinates, _shift_bboxes_cxcywhr
 
 from .abstract_obb_transform import AbstractOBBDetectionTransform
