+++ src/super_gradients/training/transforms/obb/obb_standardize.py
@@ -3,7 +3,7 @@ from typing import List, Dict
 import numpy as np
 from super_gradients.common.object_names import Processings
 from super_gradients.common.registry import register_transform
-from .obb_sample import OBBSample
+from super_gradients.training.samples.obb_sample import OBBSample
 from .abstract_obb_transform import AbstractOBBDetectionTransform
